requirements.txt
python-dotenv
pyTelegramBotAPI
requests
Flask